requirements.txt
asgiref==3.2.7
Django==2.1.15
django-cors-headers==3.2.1
django-mssql==1.8
django-pyodbc-azure==2.1.0.0
django-templated-mail==1.1.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djoser==2.0.3
mysqlclient==1.4.6
pyodbc==4.0.30
pytz==2019.3
sqlparse==0.3.1